+++ pulsar-client/src/main/java/org/apache/pulsar/client/impl/conf/ClientConfigurationData.java
@@ -63,4 +63,6 @@ public class ClientConfigurationData implements Serializable, Cloneable {
             throw new RuntimeException("Failed to clone ClientConfigurationData");
         }
     }
+
+
 }