+++ pulsar-functions/instance/src/main/java/org/apache/pulsar/functions/instance/JavaInstance.java
@@ -18,19 +18,16 @@
  */
 package org.apache.pulsar.functions.instance;
 
-import java.util.HashMap;
-import java.util.Map;
-import java.util.Optional;
-import javax.swing.text.html.Option;
 import lombok.AccessLevel;
 import lombok.Getter;
 import lombok.extern.slf4j.Slf4j;
 import org.apache.pulsar.client.api.MessageId;
 import org.apache.pulsar.client.api.PulsarClient;
 import org.apache.pulsar.functions.api.Function;
-import org.apache.pulsar.functions.proto.InstanceCommunication.MetricsData;
-import org.apache.pulsar.functions.source.PulsarSource;
+import org.apache.pulsar.functions.proto.InstanceCommunication;
 import org.apache.pulsar.io.core.Source;
+
+import org.apache.pulsar.functions.source.PulsarSource;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
@@ -41,12 +38,11 @@ import org.slf4j.LoggerFactory;
  */
 @Slf4j
 public class JavaInstance implements AutoCloseable {
-    private ContextImpl context;
 
     @Getter(AccessLevel.PACKAGE)
+    private final ContextImpl context;
     private Function function;
     private java.util.function.Function javaUtilFunction;
-    private Optional<PulsarSource> optionalPulsarSource = Optional.empty();
 
     public JavaInstance(InstanceConfig config, Object userClassObject,
                  ClassLoader clsLoader,
@@ -56,8 +52,8 @@ public class JavaInstance implements AutoCloseable {
         Logger instanceLog = LoggerFactory.getLogger("function-" + config.getFunctionDetails().getName());
 
         if (source instanceof PulsarSource) {
-            this.context = new ContextImpl(config, instanceLog, pulsarClient, clsLoader);
-            optionalPulsarSource = Optional.of((PulsarSource) source);
+            this.context = new ContextImpl(config, instanceLog, pulsarClient, clsLoader,
+                    ((PulsarSource) source).getInputConsumer());
         } else {
             this.context = null;
         }
@@ -68,17 +64,13 @@ public class JavaInstance implements AutoCloseable {
         } else {
             this.javaUtilFunction = (java.util.function.Function) userClassObject;
         }
-
     }
 
     public JavaExecutionResult handleMessage(MessageId messageId, String topicName, Object input) {
-        optionalPulsarSource.ifPresent((pulsarSource) -> {
-            this.context.setInputConsumer(pulsarSource.getConsumerForTopic(topicName));
-            this.context.setCurrentMessageContext(messageId, topicName);
-        });
-
+        if (context != null) {
+            context.setCurrentMessageContext(messageId, topicName);
+        }
         JavaExecutionResult executionResult = new JavaExecutionResult();
-
         try {
             Object output;
             if (function != null) {
@@ -93,15 +85,11 @@ public class JavaInstance implements AutoCloseable {
         return executionResult;
     }
 
-    public ContextImpl getContext() {
-        return this.context;
-    }
-
     @Override
     public void close() {
     }
 
-    public MetricsData getAndResetMetrics() {
+    public InstanceCommunication.MetricsData getAndResetMetrics() {
         return context.getAndResetMetrics();
     }
 }